+++ database/migrations/2020_07_05_163626_create_oder_details_table.php
@@ -17,9 +17,9 @@ class CreateOderDetailsTable extends Migration
             $table->id();
             $table->integer('order_id')->unsigned();
             $table->integer('product_id')->unsigned();
-            $table->float('unit_price',8,2);
+            $table->decimal('unit_price',8,2);
             $table->integer('quantity');
-            $table->float('discount',8,2);
+            $table->decimal('discount',8,2);
             $table->timestamps();
             $table->softDeletes();
             $table->foreign('order_id')->references('id')->on('orders');
